+++ tensorcircuit/backends/numpy_backend.py
@@ -164,7 +164,7 @@ class NumpyBackend(numpy_backend.NumPyBackend):  # type: ignore
     ) -> Tensor:
         return np.mean(a, axis=axis, keepdims=keepdims)
 
-    def unique_with_counts(self, a: Tensor) -> Tuple[Tensor, Tensor]:
+    def unique_with_counts(self, a: Tensor, **kws: Any) -> Tuple[Tensor, Tensor]:
         return np.unique(a, return_counts=True)  # type: ignore
 
     def min(self, a: Tensor, axis: Optional[int] = None) -> Tensor: