+++ tests/test_circuit.py
@@ -1002,10 +1002,14 @@ def test_batch_sample(backend):
     c.cnot(0, 1)
     print(c.sample())
     print(c.sample(batch=8))
-    print(c.sample(status=tc.backend.get_random_state(42)))
+    print(c.sample(random_generator=tc.backend.get_random_state(42)))
     print(c.sample(allow_state=True))
     print(c.sample(batch=8, allow_state=True))
-    print(c.sample(batch=8, allow_state=True, status=tc.backend.get_random_state(42)))
+    print(
+        c.sample(
+            batch=8, allow_state=True, random_generator=tc.backend.get_random_state(42)
+        )
+    )
 
 
 def test_expectation_y_bug():